pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hostprobe"
version = "0.1.0"
description = "Host discovery & non-resolving domain assessment tool"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "dnspython>=2.4",
    "aiohttp>=3.9",
    "python-whois>=0.9",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "aioresponses>=0.7",
]

[project.scripts]
hostprobe = "hostprobe.cli:main"

[tool.setuptools.packages.find]
include = ["hostprobe*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
